+++ packages/assignment-7/src/utils.ts
@@ -1,3 +1,20 @@
+export const getDaysInMonth = (year: number, month: number) => {
+  return new Date(year, month + 1, 0).getDate();
+};
+
+export const getWeekDates = (date: Date) => {
+  const day = date.getDay();
+  const diff = date.getDate() - day + (day === 0 ? -6 : 1);
+  const monday = new Date(date.setDate(diff));
+  const weekDates = [];
+  for (let i = 0; i < 7; i++) {
+    const nextDate = new Date(monday);
+    nextDate.setDate(monday.getDate() + i);
+    weekDates.push(nextDate);
+  }
+  return weekDates;
+};
+
 export const formatWeek = (date: Date): string => {
   const year = date.getFullYear();
   const month = date.getMonth() + 1;
